Add initial implementation of post rename
Initial shot at supporting a "post exchange" rename of oldfile. This
allows for the following pattern:

  Swap    `/tmp/foo` with `/tmp/bar`
  Rename  `/tmp/bar` to `/tmp/oldbar`

The post exchange rename will always happen in the same directory as the
old file. Meaning the following is not currently supported:

  Rename  `/tmp/bar` to `/tmp/something/bar`

Since the rename happens in a second, separate atomic transaction
anyhow, its likely not worth the effort to support this case.  The user
instead should just run a separate `mv` or `xchg` command.